requirements.txt
openai-whisper==20231117
pydub==0.25.1
soundfile==0.12.1
torch==2.2.1
numpy==1.26.4 